cast double to long

+++ android/src/main/java/com/ocetnik/timer/BackgroundTimerModule.java
@@ -86,7 +86,7 @@ public class BackgroundTimerModule extends ReactContextBaseJavaModule {
                         .emit("backgroundTimer.timeout", id);
                 }
            }
-        }, timeout);
+        }, (long) timeout);
     }
 
     /*@ReactMethod
